requirements.txt
# packages
ipykernel
pandas==2.1.1
sqlalchemy==2.0.23
python-dotenv==1.0.0
psycopg2-binary==2.9.9
numpy==1.26.0
pytest==7.4.3